This has us completing one more point of our todo checklist.[General]: Most (hopefully all) time/frame related engine side 32 bit types are properly converted to 64 bit types. [Net]: Wire serverframe appropriately using WriteIntBase128/ReadIntBase128
This means one more checked point on our todo list.

+++ inc/common/cmd.h
@@ -48,8 +48,8 @@ typedef struct cmdbuf_s {
     char        *text; // may not be NULL terminated
     size_t      cursize;
     size_t      maxsize;
-    int         waitCount;
-    int         aliasCount; // for detecting runaway loops
+    int64_t     waitCount;
+	int64_t		aliasCount; // for detecting runaway loops
     void        (*exec)(struct cmdbuf_s *, const char *);
 } cmdbuf_t;
